feature: Adicionado verificacao se existe transacoes vinculadas a catergoria antes de excluir.

+++ src/services/category.service.js
@@ -37,6 +37,12 @@ const listCategoryById = async (id) => {
 
 const deleteCategory = async (id) => {
   try {
+    const existTransactionsInCategory = await categoryRepository.existTransactionsInCategory(id);
+
+    if (existTransactionsInCategory) {
+      throw new AppError("Não é possível excluir uma categoria que possui transações associadas.");
+    }
+
     const category = await categoryRepository.deleteCategory(id);
     return category;
   } catch (error) {
